feat(dropdown): implement dropdown component with type-safe props and selection handling
 Created DropDown.types.ts to define DropDownOption, SelectedOption, and DropdownProps interfaces
- Implemented DropDown.tsx component using internal isOpen state and external selection state via props
- Created DropDownPage.tsx to demonstrate usage with selection state and handleSelect callback
