+++ routing-system/src/types/DropDown.types.ts
@@ -1,0 +1,12 @@
+export interface DropDownOption {
+  label: string;
+  value: string;
+}
+
+export interface DropdownProps {
+  options: DropDownOption[];
+  selection: SelectedOption;
+  onSelect: (option: DropDownOption) => void;
+}
+
+export type SelectedOption = DropDownOption | null;
